Move exceptions to BatchException
