+++ src/HttpPsrClient.php
@@ -2,6 +2,7 @@
 
 namespace Http\Client;
 
+use Http\Client\Exception\BatchException;
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
 
@@ -30,9 +31,10 @@ interface HttpPsrClient
      * @param RequestInterface[] $requests
      * @param array              $options
      *
-     * @return ResponseInterface[]
+     * @return BatchResult
      *
      * @throws Exception
+     * @throws BatchException
      */
     public function sendRequests(array $requests, array $options = []);
 }
